Removed shared dependency
