+++ LocalyticsXamarin/LocalyticsXamarin.Shared/SharedInAppCampaign.cs
@@ -1,0 +1,4 @@
+﻿namespace LocalyticsXamarin.Shared
+{
+    public abstract class SharedInAppCampaign {}
+}
