+++ LocalyticsXamarin/LocalyticsXamarin.Shared/SharedInboxCampaign.cs
@@ -1,0 +1,4 @@
+﻿namespace LocalyticsXamarin.Shared
+{
+    public abstract class SharedInboxCampaign { }
+}
